successfully tested polyfit using qr decomp
